jwt delete

+++ controllers/user.js
@@ -54,8 +54,6 @@ const updateUser = async (req, res) => {
       console.log("logged on userane: ", loggedUsername);
       console.log("actual useranme: ", username);
 
-
-
       // Check if the user is authorized to perform the update
       if (username !== loggedUsername) {
          return res.status(403).json({ errors: ['User is not authorized to update this profile'] });
@@ -78,17 +76,36 @@ const updateUser = async (req, res) => {
 };
 
 
+   const deleteUser = async (req, res) => {
+      try {
+         const username = req.params.id; // Get username from request parameters
+   
+         // Extract the token from the request headers
+         const token = req.headers.authorization.split(' ')[1];
+   
+         // Verify the token using the token service
+         const decodedToken = tokenService.verifyToken(token);
+         const loggedUsername = decodedToken.username;
+   
+         console.log("logged on username: ", loggedUsername);
+         console.log("actual username: ", username);
+   
+         // Check if the user is authorized to delete the profile
+         if (username !== loggedUsername) {
+            return res.status(403).json({ errors: ['User is not authorized to delete this profile'] });
+         }
+   
+         const user = await UserService.deleteUser(username);
+         if (!user) {
+            return res.status(404).json({ errors: ['User not found'] });
+         }
+         res.json(user);
+      } catch (error) {
+         console.error("Error deleting user:", error);
+         res.status(500).json({ errors: ['Failed to delete user'] });
+      }
+   };
 
-const deleteUser = async (req, res) => {
-   const username = req.params.id; // Get username from request parameters
-   const user = await UserService.deleteUser(username);
-   if (!user) {
-      return res.status(404).json({ errors: ['User not found'] });
-      // add try and catch!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   }
-   res.json(user);
-
-};
